genai-rag-web-ingestion/api/requirement.txt
fastapi
uvicorn
azure-search-documents
openai
python-dotenv
azure.core.credentials
